+++ hourly.sh
@@ -39,9 +39,24 @@ sleep 1h
 kill `cat ${WORK_HOUR}/recording.pid`
 rm ${WORK_HOUR}/recording.pid
 # record weather data
-curl "wttr.in/38.103516,-121.288475?format=insert+into+weather+values+('%C','%t','%h','%w','%p','%P',datetime('now'));" > ${BARCHART_HOME}/weather.sql
-sqlite3 ${BARCHART_HOME}/birds.db < ${BARCHART_HOME}/weather.sql
-rm ${BARCHART_HOME}/weather.sql
+MAXTRYS=5
+for i in $(seq 1 $MAXTRYS)
+do
+	curl  -H "Cache-Control: no-cache, no-store" "wttr.in/38.103516,-121.288475?format=insert+into+weather+values+('%C','%t','%h','%w','%p','%P',datetime('now','localtime'));\n" > ${BARCHART_HOME}/weather.sql
+	EXITCODE=$?
+	if [[ $EXITCODE -eq 0 ]]
+		then
+		sqlite3 ${BARCHART_HOME}/birds.db < ${BARCHART_HOME}/weather.sql
+		echo "weather sucess on attempt ${i}"
+		rm ${BARCHART_HOME}/weather.sql
+		break   
+	else
+		echo "curl: ${EXITCODE}"
+		cat ${BARCHART_HOME}/weather.sql
+		echo "===== weather FAILURE ====="
+		sleep 5
+	fi
+done
 # directory for audio files and analyze each
 # BirdNET-Analyzer using the species_list.txt created by weekly.sh, creates the CSV file
 pushd ${ANALYZER_HOME}
